another commit

+++ controllers/UsersController.js
@@ -3,15 +3,13 @@ const jwt = require('jsonwebtoken');
 const secret = require('../config/keys').secretOrKey;
 const fs = require('fs-extra');
 const ObjectId = require('mongoose').Types.ObjectId; 
-
 const UserModel = require('../models/User');
 const UserRolesModel = require('../models/UserRole');
-
 const validateRegisterInput = require('../validation/register');
 const validateLoginInput = require('../validation/login');
 const validatePasswords = require('../validation/checkpassword');
-
 const sendActivationEmail = require('../controllers/EmailController').sendActivationEmail;
+const getTokenEXP = require('../helpers/getTokenEXP').getDecodedTokenEXP;
 
 //catch 
 exports.registerUser = async (req, res) => {
@@ -29,7 +27,7 @@ exports.registerUser = async (req, res) => {
             return res.json(errors);
         }
 
-        const token = await jwt.sign({email: req.body.email},secret,{ expiresIn: 86400 });
+        const jwtToken = await jwt.sign({email: req.body.email},secret,{ expiresIn: 86400 });
 
         const userRoles = await UserRolesModel.findOne({name: 'User'});
 
@@ -42,27 +40,26 @@ exports.registerUser = async (req, res) => {
             username: req.body.username,
             password: hash,
             emailConfirmation: {
-                token: 'Bearer ' + token,
-                tokenExparation: Date.now() + 86400000,
+                token: 'Bearer ' + jwtToken,
+                tokenExparation: new Date(getTokenEXP(jwtToken)),
                 confirmed: false
             }
         });
 
         await addNewUser.save();
 
-        sendActivationEmail(addNewUser, `Bearer ${token}`);
+        sendActivationEmail(addNewUser, `Bearer ${jwtToken}`);
 
         return res.json({ 
             message: `Please ${addNewUser.username} go to your Email and confirm your email address in order to log in.`, 
             addNewUser,
-            token: 'Bearer ' + token
+            token: 'Bearer ' + jwtToken
         })
     } catch(err) {
         return res.json({ failedMessage: err.message });
     }
 
 }
-
 // catch using $or
 exports.loginUser = async (req, res) => {
 
@@ -105,77 +102,48 @@ exports.loginUser = async (req, res) => {
         return res.json({ failedMessage: err.message });
     }
 }
-
 // catch
 exports.resetPassword = async (req, res) => {
     const {errors, isValid} = validatePasswords(req.body);
-
     if(!isValid) { return res.json(errors); }
 
     try{
-        
         const user = await UserModel.findOne({ email: req.body.email });
-    
         const expiresIn = new Date(user.resetPassword.expiresIn).getTime();
-    
+
         if(expiresIn < Date.now()) {
             errors.errors.expiredToken = 'Token has expired please make request again !!';
-    
             return res.json(errors);
         }
 
         const byCrypt = await bcrypt.compare(req.body.password, user.password);
 
-        if(byCrypt) {
-            return res.json({ successMessage: 'Provided password is already in use'});
-        }
+        if(byCrypt) { return res.json({ successMessage: 'Provided password is already in use'}); }
     
-        bcrypt.genSalt(10, (err, salt) => {
-            bcrypt.hash(req.body.password, salt, async (err,hash) => {
-                if(err) throw err;
-    
-                await UserModel.updateOne({email: req.body.email}, {
-                    resetPassword: {},
-                    password: hash
-                });
-    
-                return res.json({ successMessage: 'Password has bees successfuly changed'});
-            });
-        });
+        const hash = await bcrypt.hash(req.body.password,10);
 
+        await UserModel.updateOne({email: req.body.email}, { resetPassword: {}, password: hash });
+
+        return res.json({ successMessage: 'Password has bees successfuly changed'});
     }catch(err) {
         return res.json({ failedMessage: err.message });
     }
 
 }
-
 // catch
 exports.updateProfilePicture = async (req, res) => {
     try{
-
-        if(!req.file){
-            return res.json({ failedMessage: 'Attached file is not an image.' });
-        }
+        if(!req.file){ return res.json({ failedMessage: 'Attached file is not an image.' }); }
 
         const user = await UserModel.findOne({username: req.body.username});
 
-        if(user.profilePicture === req.file.path) {
-            return res.json({ failedMessage: 'Selected picture is same as the one u already use !' });
-        }
+        if(user.profilePicture === req.file.path) { return res.json({ failedMessage: 'Selected picture is same as the one u already use !' }); }
+        if(user.profilePicture !== ''){ await fs.remove(`${user.profilePicture}`) }
 
-        if(user.profilePicture !== ''){
-            await fs.remove(`${user.profilePicture}`)
-        }
-
-
-        await UserModel.updateOne({username: req.body.username}, {
-            profilePicture: req.file.path
-        });
+        await UserModel.updateOne({username: req.body.username}, { profilePicture: req.file.path });
 
         const updatedUser = await UserModel.findOne({username: req.body.username});
-
         const role = await UserRolesModel.findOne({ _id: updatedUser.role });
-
         const expiresIn = req.body.rememberMe === true ? null : { expiresIn: 3600 };
 
         const userInfo = {
@@ -190,11 +158,7 @@ exports.updateProfilePicture = async (req, res) => {
 
         const jwtToken = await jwt.sign(userInfo,secret,expiresIn);
 
-        return res.json({
-            token: 'Bearer ' + jwtToken,
-            user: userInfo,
-            successMessage: 'Profile Picture is successfuly updated !'
-        });
+        return res.json({ token: 'Bearer ' + jwtToken, user: userInfo, successMessage: 'Profile Picture is successfuly updated !' });
 
     }catch (err) {
         return res.json({ failedMessage: err.message })
@@ -203,33 +167,26 @@ exports.updateProfilePicture = async (req, res) => {
 
 exports.updatePassword = async (req, res) => {
     const {errors, isValid} = validatePasswords(req.body);
-
     if(!isValid) { return res.json(errors); }
 
     try{
-
         const user = await UserModel.findOne({ username: req.body.username });
-
         const byCrypt = await bcrypt.compare(req.body.password, user.password);
 
-        if(byCrypt) {
-            return res.json({ failedMessage: 'Provided password is already in use'});
-        }
+        if(byCrypt) { return res.json({ failedMessage: 'Provided password is already in use'}); }
 
         const hashPwd = await bcrypt.hash(req.body.password,10);
         await UserModel.updateOne({username: req.body.username}, {password: hashPwd});
-        return res.json({successMessage: 'Password has bees successfuly changed!'});
 
+        return res.json({successMessage: 'Password has bees successfuly changed!'});
     }catch(err) {
         return res.json({ failedMessage: err.message });
     }
-
 }
 
 exports.getAllUsers = async (req, res) => {
     try{
         const users = await UserModel.find().select('name _id username email profilePicture');
-
         return res.json(users);
     }catch (err){
         console.log(err)
@@ -238,20 +195,11 @@ exports.getAllUsers = async (req, res) => {
 
 exports.getSingleUser = async (req, res) => {
     try{
-
         const user = await UserModel.findOne({_id: new ObjectId(req.query.id)}).select('profilePicture name email role username userInfo -_id')
-
         const role = await UserRolesModel.findOne({ _id: user.role }).select('isAdmin name permissions -_id');
-
-        const payload = {
-            ...user._doc,
-            role: {
-                ...role._doc
-            }
-        }
+        const payload = { ...user._doc, role: { ...role._doc } }
 
         return res.json(payload)
-
     }catch(err) {
         console.log(err)
     }
@@ -260,76 +208,45 @@ exports.getSingleUser = async (req, res) => {
 exports.deleteUser = async (req, res) => {
     try{
         const user = await UserModel.deleteOne({ _id: new ObjectId(req.query.id) });
-
         return res.json({ successMessage: 'User has been deleted !!!' })
-
     }catch(err) {
         console.log(err);
     }
 }
 
 exports.addEmployeeUser = async (req, res) => {
-
     const {errors, isValid} = validateRegisterInput(req.body);
-    
-    if(!isValid) {
-        return res.json(errors);
-    }
+    if(!isValid) return res.json(errors);
 
     try{
+        const user = await UserModel.find({$or:[{ email: req.body.email}, {username: req.body.username }]})
 
-        let user = await UserModel.findOne({ email: req.body.email });
-
-        if(user) {
-            errors.errors.email = 'Email already exists';
+        if(user.length > 0) {
+            (user[0].email || user[1].email) === req.body.email ? errors.errors.email = 'Email already in use !' : null;
+            (user[0].username || user[1].username) === req.body.username ? errors.errors.username = 'Username already in use!' : null;
             return res.json(errors);
         }
-
-        user = await UserModel.findOne({ username: req.body.username });
-
-        if(user) {
-            errors.errors.username = 'Username is already taken';
-            return res.json(errors);
-        }
-
-        const token = await jwt.sign({email: req.body.email},secret,{ expiresIn: 86400 });
 
         const userRoles = await UserRolesModel.findOne({name: 'User'});
-        
-        bcrypt.genSalt(10, (err, salt) => {
-            bcrypt.hash(req.body.password, salt, async (err,hash) => {
-                if(err) throw err;
-                const addNewUser = new UserModel({
-                    name: req.body.name,
-                    email: req.body.email,
-                    role: userRoles._id,
-                    username: req.body.username,
-                    password: hash,
-                    emailConfirmation: {
-                        token: 'Bearer ' + token,
-                        tokenExparation: Date.now() + 86400000,
-                        confirmed: false
-                    }
-                });
+        const hash = await bcrypt.hash(req.body.password,10);
 
-                try{
-                    await addNewUser.save();
-
-                    sendActivationEmail(addNewUser.email, 'Bearer ' + token);
-
-                    return res.json({ 
-                        message: `Please ${addNewUser.username} go to your Email and confirm your email address in order to log in.`, 
-                        addNewUser,
-                        token: 'Bearer ' + token
-                    })
-                }catch(err){
-                    console.log(err)
-                }
-            });
+        const addNewUser = new UserModel({
+            name: req.body.name,
+            email: req.body.email,
+            role: userRoles._id,
+            username: req.body.username,
+            password: hash,
+            emailConfirmation: {
+                token: "",
+                tokenExparation: null,
+                confirmed: true
+            }
         });
+
+        await addNewUser.save();
+        return res.json({ message: `Please ${addNewUser.username} go to your Email and confirm your email address in order to log in.` });
     } catch(err) {
         errors.errors.email = err;
-        return res.json(errors);
+        return res.json({ failedMessage: err.message });
     }
-
 }
